Removes the timestamp part of the abigen comment
The abigen adds a comment to the generated ABI file. This comment
includes a timestamp information at which time the ABI file was generated.
This is problematic when trying to reproduce an exact identical ABI file
since the timestamp will make the file looks different.

This PR removes the actual timestamp information from the comment and
only keep the static text instead.

By not having the timestamp, someone can checkout a particular tag,
perform the build step and validate that the sha256 of the ABI file has
not changed.